Split Rhythm Tweezers, Fireworks and The Clappy Trio game engines into their own C files.
