+++ src/engines/clappy_trio.c
@@ -1,0 +1,55 @@
+#include "src/engines/clappy_trio.h"
+
+asm(".include \"include/gba.inc\""); // Temporary
+
+
+/* THE CLAPPY TRIO */
+
+
+#include "asm/scenes/clappy_trio/asm_08030384.s"
+
+#include "asm/scenes/clappy_trio/asm_080303a4.s"
+
+#include "asm/scenes/clappy_trio/asm_0803050c.s"
+
+#include "asm/scenes/clappy_trio/asm_0803051c.s"
+
+#include "asm/scenes/clappy_trio/asm_0803055c.s"
+
+#include "asm/scenes/clappy_trio/asm_08030588.s"
+
+#include "asm/scenes/clappy_trio/asm_0803068c.s"
+
+#include "asm/scenes/clappy_trio/asm_0803074c.s"
+
+#include "asm/scenes/clappy_trio/asm_0803080c.s"
+
+#include "asm/scenes/clappy_trio/asm_0803088c.s"
+
+#include "asm/scenes/clappy_trio/asm_08030898.s"
+
+#include "asm/scenes/clappy_trio/asm_080308a4.s"
+
+#include "asm/scenes/clappy_trio/asm_080308b8.s"
+
+#include "asm/scenes/clappy_trio/asm_080308bc.s"
+
+#include "asm/scenes/clappy_trio/asm_080308d4.s"
+
+#include "asm/scenes/clappy_trio/asm_080308f0.s"
+
+#include "asm/scenes/clappy_trio/asm_080308f4.s"
+
+#include "asm/scenes/clappy_trio/asm_080309a8.s"
+
+#include "asm/scenes/clappy_trio/asm_080309f4.s"
+
+#include "asm/scenes/clappy_trio/asm_08030a10.s"
+
+#include "asm/scenes/clappy_trio/asm_08030a60.s"
+
+#include "asm/scenes/clappy_trio/asm_08030bd8.s"
+
+#include "asm/scenes/clappy_trio/asm_08030bf0.s"
+
+#include "asm/scenes/clappy_trio/asm_08030c14.s"
