+++ src/engines/clappy_trio.h
@@ -1,0 +1,42 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+// Engine Macros/Enums:
+
+// OAM Animations:
+
+// Palettes:
+
+// Sound Effects:
+
+// Engine Data:
+
+// Engine Definition Data:
+
+// Functions:
+// extern ? func_08030384(?);
+// extern ? func_080303a4(?);
+// extern ? func_0803050c(?);
+// extern ? func_0803051c(?);
+// extern ? func_0803055c(?);
+// extern ? func_08030588(?);
+// extern ? func_0803068c(?);
+// extern ? func_0803074c(?);
+// extern ? func_0803080c(?);
+// extern ? func_0803088c(?);
+// extern ? func_08030898(?);
+// extern ? func_080308a4(?);
+// extern ? func_080308b8(?);
+// extern ? func_080308bc(?);
+// extern ? func_080308d4(?);
+// extern ? func_080308f0(?);
+// extern ? func_080308f4(?);
+// extern ? func_080309a8(?);
+// extern ? func_080309f4(?);
+// extern ? func_08030a10(?);
+// extern ? func_08030a60(?);
+// extern ? func_08030bd8(?);
+// extern ? func_08030bf0(?);
+// extern ? func_08030c14(?);
