+++ src/engines/fireworks.h
@@ -1,0 +1,39 @@
+#pragma once
+
+#include "global.h"
+#include "sound.h"
+
+// Engine Macros/Enums:
+
+// OAM Animations:
+
+// Palettes:
+
+// Sound Effects:
+
+// Engine Data:
+
+// Engine Definition Data:
+
+// Functions:
+// extern ? func_0802f3a4(?); // [func_0802f3a4] GFX_INIT Func_02
+// extern ? func_0802f3b4(?); // [func_0802f3b4] GFX_INIT Func_01
+// extern ? func_0802f3f4(?); // [func_0802f3f4] GFX_INIT Func_00
+// extern ? func_0802f420(?); // [func_0802f420] MAIN - Init
+// extern ? func_0802f5b8(?); // [func_0802f5b8] ENGINE Func_00 - Set Pattern Mode
+// extern ? func_0802f5f0(?); // [func_0802f5f0] ENGINE Func_01 - Play Sound
+// extern ? func_0802f650(?); // [func_0802f650] ENGINE Func_02 - Set Pattern
+// extern ? func_0802f664(?); // [func_0802f664] Update Particles
+// extern ? func_0802f74c(?); // [func_0802f74c] Create Explosion
+// extern ? func_0802fc38(?); // [func_0802fc38] MAIN - Update
+// extern ? func_0802fc6c(?); // [func_0802fc6c] MAIN - Close (STUB)
+// extern ? func_0802fc70(?); // [func_0802fc70] CUE - Spawn
+// extern ? func_0802ff70(?); // [func_0802ff70] CUE - Update
+// extern ? func_08030114(?); // [func_08030114] CUE - Despawn
+// extern ? func_0803012c(?); // [func_0803012c] CUE - Hit
+// extern ? func_080301d0(?); // [func_080301d0] CUE - Barely
+// extern ? func_08030288(?); // [func_08030288] CUE - Miss
+// extern ? func_08030294(?); // [func_08030294] MAIN - Input Event (STUB)
+// extern ? func_08030298(?); // [func_08030298] COMMON Func_00 - STUB
+// extern ? func_0803029c(?); // [func_0803029c] COMMON Func_01 - Display Text
+// extern ? func_0803031c(?); // [func_0803031c] COMMON Func_02 - Set Tutorial Mode
