+++ src/scenes.h
@@ -3,102 +3,6 @@
 #include "global.h"
 #include "sound.h"
 
-// Rhythm Tweezers
-
-#define RT_VEGETABLE_BG_MAP_L *(u32 *)(VRAMBase + 0xF000) // VRAM BG Map for vegetable textures (left).
-#define RT_VEGETABLE_BG_MAP_R *(u32 *)(VRAMBase + 0xF800) // VRAM BG Map for vegetable textures (right).
-
-extern void func_0802e750(void);        // [func_0802e750] Initialise Vegetable Face
-extern void func_0802e828(u32);         // [func_0802e828] ENGINE Func_02 - Scroll to New Vegetable
-extern void func_0802e89c(void);        // [func_0802e89c] ENGINE Func_04 - Play Cash Register SFX
-extern void func_0802e8ac(void);        // [func_0802e8ac] Update Vegetable (Scrolling)
-extern void func_0802e96c(u32);         // [func_0802e96c] ENGINE Func_03 - Define New Vegetable Type
-extern void func_0802e97c_stub(void);   // [func_0802e97c] ENGINE Func_06 - Unknown (STUB)
-extern void func_0802e980(void);        // [func_0802e980] Update Vegetable
-extern void func_0802e99c(void);        // [func_0802e99c] Initialise Falling Hairs
-extern void func_0802ea20(void);        // [func_0802ea20] Update Falling Hairs
-extern void func_0802ea74(u32);         // [func_0802ea74] Create Falling Hair
-extern void func_0802eaf8(void);        // [func_0802eaf8] Initialise Tweezers
-extern void func_0802eb7c(void);        // [func_0802eb7c] ENGINE Func_00 - Spawn Tweezers
-extern void func_0802eba0(void);        // [func_0802eba0] Update Tweezers (Position)
-extern void func_0802ebdc(void);        // [func_0802ebdc] Update Tweezers (Vertical Screen Scroll)
-extern void func_0802ebf8(void);        // [func_0802ebf8] Update Tweezers
-extern void func_0802ec50(void);        // [func_0802ec50] GFX_INIT Func_02
-extern void func_0802ec60(void);        // [func_0802ec60] GFX_INIT Func_01
-extern void func_0802eca0(void);        // [func_0802eca0] GFX_INIT Func_00
-extern void func_0802eccc(u8);          // [func_0802eccc] MAIN - Init
-extern void func_0802edc4_stub(void);   // [func_0802edc4] ENGINE Func_09 - STUB
-extern void func_0802edc8(void);        // [func_0802edc8] Update Mask
-extern void func_0802ee00(void);        // [func_0802ee00] ENGINE Func_07 - Show Mask
-extern void func_0802ee10(void);        // [func_0802ee10] ENGINE Func_08 - Hide Mask
-extern void func_0802ee24(void);        // [func_0802ee24] MAIN - Update
-extern void func_0802ee40_stub(void);   // [func_0802ee40] MAIN - Close (STUB)
-extern void func_0802ee44(void);        // [func_0802ee44] ENGINE Func_01 - Reset Hair Placement Cycle
-extern void func_0802ee6c(void);        // [func_0802ee6c] Update Hair Placement Cycle
-extern void func_0802ee7c(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802ee7c] CUE - Spawn
-extern u32  func_0802ef54(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802ef54] CUE - Update (Short Hair)
-extern u32  func_0802ef68(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802ef68] CUE - Update (Long Hair)
-extern void func_0802f164(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802f164] CUE - Despawn
-extern void func_0802f170(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802f170] CUE - Hit (Short Hair)
-extern void func_0802f21c(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802f21c] CUE - Hit/Barely (Long Hair)
-extern void func_0802f2a0(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802f2a0] CUE - Barely (Short Hair)
-extern void func_0802f330(u32, struct RhythmTweezersCue *, u32, u32);   // [func_0802f330] CUE - Miss
-extern void func_0802f33c(void);        // [func_0802f33c] MAIN - Input Event
-extern void func_0802f378_stub(void);   // [func_0802f378] COMMON Func_00 - STUB
-extern void func_0802f37c_stub(void);   // [func_0802f37c] COMMON Func_01 - STUB
-extern void func_0802f380(void);        // [func_0802f380] ENGINE Func_05 - Unknown (Unused)
-
-// Fireworks
-
-// extern ? func_0802f3a4(?); // [func_0802f3a4] GFX_INIT Func_02
-// extern ? func_0802f3b4(?); // [func_0802f3b4] GFX_INIT Func_01
-// extern ? func_0802f3f4(?); // [func_0802f3f4] GFX_INIT Func_00
-// extern ? func_0802f420(?); // [func_0802f420] MAIN - Init
-// extern ? func_0802f5b8(?); // [func_0802f5b8] ENGINE Func_00 - Set Pattern Mode
-// extern ? func_0802f5f0(?); // [func_0802f5f0] ENGINE Func_01 - Play Sound
-// extern ? func_0802f650(?); // [func_0802f650] ENGINE Func_02 - Set Pattern
-// extern ? func_0802f664(?); // [func_0802f664] Update Particles
-// extern ? func_0802f74c(?); // [func_0802f74c] Create Explosion
-// extern ? func_0802fc38(?); // [func_0802fc38] MAIN - Update
-// extern ? func_0802fc6c(?); // [func_0802fc6c] MAIN - Close (STUB)
-// extern ? func_0802fc70(?); // [func_0802fc70] CUE - Spawn
-// extern ? func_0802ff70(?); // [func_0802ff70] CUE - Update
-// extern ? func_08030114(?); // [func_08030114] CUE - Despawn
-// extern ? func_0803012c(?); // [func_0803012c] CUE - Hit
-// extern ? func_080301d0(?); // [func_080301d0] CUE - Barely
-// extern ? func_08030288(?); // [func_08030288] CUE - Miss
-// extern ? func_08030294(?); // [func_08030294] MAIN - Input Event (STUB)
-// extern ? func_08030298(?); // [func_08030298] COMMON Func_00 - STUB
-// extern ? func_0803029c(?); // [func_0803029c] COMMON Func_01 - Display Text
-// extern ? func_0803031c(?); // [func_0803031c] COMMON Func_02 - Set Tutorial Mode
-
-// Clappy Trio
-
-// extern ? func_08030384(?);
-// extern ? func_080303a4(?);
-// extern ? func_0803050c(?);
-// extern ? func_0803051c(?);
-// extern ? func_0803055c(?);
-// extern ? func_08030588(?);
-// extern ? func_0803068c(?);
-// extern ? func_0803074c(?);
-// extern ? func_0803080c(?);
-// extern ? func_0803088c(?);
-// extern ? func_08030898(?);
-// extern ? func_080308a4(?);
-// extern ? func_080308b8(?);
-// extern ? func_080308bc(?);
-// extern ? func_080308d4(?);
-// extern ? func_080308f0(?);
-// extern ? func_080308f4(?);
-// extern ? func_080309a8(?);
-// extern ? func_080309f4(?);
-// extern ? func_08030a10(?);
-// extern ? func_08030a60(?);
-// extern ? func_08030bd8(?);
-// extern ? func_08030bf0(?);
-// extern ? func_08030c14(?);
-
 // Samurai Slice
 
 // extern ? func_08030c48(?);
