--- src/scenes/fireworks.inc.c
@@ -1,41 +1,0 @@
-#include "asm/scenes/fireworks/asm_0802f3a4.s"
-
-#include "asm/scenes/fireworks/asm_0802f3b4.s"
-
-#include "asm/scenes/fireworks/asm_0802f3f4.s"
-
-#include "asm/scenes/fireworks/asm_0802f420.s"
-
-#include "asm/scenes/fireworks/asm_0802f5b8.s"
-
-#include "asm/scenes/fireworks/asm_0802f5f0.s"
-
-#include "asm/scenes/fireworks/asm_0802f650.s"
-
-#include "asm/scenes/fireworks/asm_0802f664.s"
-
-#include "asm/scenes/fireworks/asm_0802f74c.s"
-
-#include "asm/scenes/fireworks/asm_0802fc38.s"
-
-#include "asm/scenes/fireworks/asm_0802fc6c.s"
-
-#include "asm/scenes/fireworks/asm_0802fc70.s"
-
-#include "asm/scenes/fireworks/asm_0802ff70.s"
-
-#include "asm/scenes/fireworks/asm_08030114.s"
-
-#include "asm/scenes/fireworks/asm_0803012c.s"
-
-#include "asm/scenes/fireworks/asm_080301d0.s"
-
-#include "asm/scenes/fireworks/asm_08030288.s"
-
-#include "asm/scenes/fireworks/asm_08030294.s"
-
-#include "asm/scenes/fireworks/asm_08030298.s"
-
-#include "asm/scenes/fireworks/asm_0803029c.s"
-
-#include "asm/scenes/fireworks/asm_0803031c.s"
